VLab Issue #16259 Event bus fixed to avoid out-of-order messages
Fixed the event bus so that it does not occasionally deliver
asynchronously-published messages in a different order from that in
which they were published. This bug occurred because of the use of
multiple dispatch and handler threads within the event bus. Also
removed any thread-count-specifying code from app builder and into
the event bus, as the app builder doesn't need to know about this.

Change-Id: I5aa118cb362705e22b94ce1e5291f11da367ccb6