VLab Issue #15934 Infrastructure to support curved geometries
Created new IAdvancedGeometry interface with three concrete
implementations for the moment (Ellipse, GeometryWrapper (wrapping a JTS
Geometry), and AdvancedGeometryCollection). This hierarchy of geometries
allows for curves, which JTS/Shapely geometries do not allow.

Replaced use of JTS/Shapely geometries in hazard events and visual
features with the new advanced geometries. This involved extensive work
throughout the codebase, as the old geometries were used nearly
everywhere.

This changeset does not add the GUI elements (button in toolbar, and
mouse interaction) necessary to actually draw an Ellipse, but gets us
98% of the way there. The purpose of this code review is to ensure that
the extensive changes in support of the new geometries have not broken
anything.

Change-Id: I6791c5f03a1961cd879ca423d50cc45be03f20f1