VLab Issue #22736 Interdependency scripts run when event is selected
Changed HID behavior so that when a hazard event is selected and the HID
pops up, any interdependency script associated with the event's metadata
runs in "initialization" mode (that is, having no triggerIdentifiers
passed in as parameters). This already occurred before when the HID was
undocked (floating free of the CAVE window), but not when it was
attached to the CAVE window. This changeset ensures it is always the
case.

Also put in a fix for an unrelated bug seen once (NullPointerException)
in the drawable manager, having to do with tracking text drawables that
may be combinable with others of the same type (when text drawables
overlay one another, they are combined into one so as not to occlude
one another).