VLab Issue #22870 Frame change trigger of tools
Added support for D2D frame changes triggering tools depending upon the
configuration of EventDrivenTools.py. Changes to the frame index caused
originally by H.S. (e.g. when the user changes the selected time and the
D2D frame is changed to match it as closely as possible) do not trigger
the tools.

Fixed bug causing display to not always redraw when H.S. told it to
change frames.

Fixed bug in which the frame asked for by H.S. did not necessarily match
the H.S. selected time as well as it could have (because only the end of
the valid time range was considered when choosing the closest frame,
instead of the entire valid time range for each frame).

Fixed bug in which H.S., when the Time Match Basis in D2D, would not
reduce the number of frames when the user called for their reduction (by
changing the frame count in the D2D toolbar from 12 to 8, for example).

Change-Id: I57762ac5b310cf0a9ef6cfff63b39010c9155044