VLab Issue #22894 Remove hazard event session attributes before persist
Added configurable (via HazardTypes.py) per-hazard-type
"sessionAttributes", a list of hazard attributes that are to be removed
prior to persisting a hazard event of that type. This is being done so
as to allow recommenders, etc. to use hazard event attributes as storage
for session-specific data, but ensure that such data does not get
persisted.