VLab Issue #21873 Allow seconds-level time resolution as an option
Added the notion of "time resolution" to Hazard Services in two places:

(1) within settings (and by extension startup config), which is
    primarily used to control the granularity of the selected time
    slider in the console time line ruler; and

(2) within hazard types, which controls whether a particular hazard
    may have start and end times at second- or minute-level resolution.

Since (1) is controlled via settings, seconds-level time ticks may be
turned on or off by changing settings. The main Probabilistic setting
uses "seconds" resolution,  meaning that the current time marker on the
time ruler ticks forward approximately once every second. Other settings
use "minutes" resolution, meaning the current time ticks forward about
once a minute, just as before.

Hazard events with seconds-level resolution are allowed to have start
and end times that do not fall on minute boundaries; this means the user
can drag their start and end time thumbs to between-minute-boundary
positions.

Altered time-manipulating megawidgets to allow the specification down to
the second, not just to the minute, if configured to do so by their
specifiers.

Visual features are now allowed to have time boundaries for their
properties that do not lie on minute boundaries. Also, the upper
boundary for visual features' properties is now exclusive, that is,
visual feature properties are considered valid with respect to time if
the selected time falls at or beyond their lower time boundary, but
below (not on) their upper time boundary. The visual feature creation
being done by various recommenders and so on has been altered to work
with this new scheme.

Changed mouse-wheel-based zooming of the time line ruler to zoom in
centered on the point over which the mouse cursor is hovering at that
moment, instead of zooming from the center of the time line. This makes
such zooming a lot more convenient when the user is trying to zoom in
on a particular point in time.

Fixed null pointer exceptions that occurred very occasionally when
switching perspectives.

Fixed null pointer exceptions in HID, Spatial Display, and session event
manager revealed during testing with Tiffany and Greg on the 21st.

Change-Id: I70b4a5b67fe4864304ea9a03d04e7a43b96ff301