VLab Issue #26330
Provides functionality to create/issue International SIGMETs. There are several tool-related files in order to satisfy the requirements
as determined by the MWOs. The CreateVALayerForecastTool created Visual Features of the current and future position of the hazard
objects, and can accomodate up to 3 forecast layers. The InternationalSigmetTool serves as the update for the visualFeatures, and
will be used for any additional development tools in the future. This follows the paradigm used with the convective SIGMET and the
LineAndPointTool.

The PopulateTCMetaDataTool is used to read through the TCA files generated by the NHC and CPHC, in order to populate the underlying
megawidget and metadata for tropical cyclone SIGMETs.

The rest of the files are the standard entries to add the product, as well as the metadata (megawidget), product generator, and formatters.

Change-Id: Id1c8b3f3788e98c9d53d7d8bccb9d54446a3ab1d