VLab Issue #26855 Wrapping long labels in Product Editor
Changed the Product Editor to use wrapping labels in its data editor to
label its editable fields. This is required in order to allow for long
labels that do not cause the scrollable area to expand horizontally to
the point where a horizontal scrollbar is required (which makes the
Product Editor unwieldy to use in such cases).
    
Also added extensive logic to the code that handles resizes of parents
or other ancestors of wrapping Label megawidgets, so that the Labels set
their widths more intelligently in response to such resizes. This fixed
some bugs discovered in the course of creating this changeset.