VLab Issue #15556 Console Refactor and Visible History List
Refactored the console in order to untangle the various modules and
decompose it into parts with far fewer interconnections, adhering to the
H.S. MVP design.

Added the ability to show the history lists of events in the console.

Added a "visible in history list" flag that can be set to true or false
each time a hazard event is saved to the database. Any manually saved
events are flagged as visible, as are those resulting from the issuance
of an event. Saves to the database prompted by a recommender may result
in visible or invisible entries in the history list, as appropriate to
the particular recommender.

Added the viewing in the HID of historical snapshots of events that the
user has selected (in the history lists of events in the console). HID
tabs created for such historical snapshots use a italicized font, as
well as a dark gray text color instead of black, in the tab header, in
order to differentiate them from the current versions of events. They
also provide a tooltip when the user hovers over the tab header that
indicates they are historical, and the timestamp indicating when they
were persisted to the database. All GUI elements for them (start/end
time, category, type, and metadata) is rendered uneditable, as being
historical snapshots, they can never be changed by the user.

Added code to the Swath Recommender to turn off visibility of events
saved to the database unless the user has just altered the automation
level of an event from automated to something else.

Fixed a couple of bugs in spatial display causing visual representations
of hazard events to appear and disappear at odd times.

Removed ugly multiple redraw of console tree contents when sorting.

Fixed bug from last patchset that caused an exception to be thrown if an
event was added after sorting in the console.

Fixed misalignment of the layout of the timeline ruler in the header of
the time range column in the console tree when the tree's vertical
scrollbar changed visibility, and when columns were added or removed.

Made the Probabilistic Convective settings "Event ID" column wider by
default, so that the entire event identifier can be displayed for
historical entries in the console table.

Ensured that reverting a hazard event's status from "ending" to "issued"
does not result in a new copy of the event being added to the history
list.

Disabled time sliders in console temporarily (will be fixed with
per-hazard-type solution in issue #26716).

Fixed bug that caused events that had their status change to one that
was invisible according to the filters for the current settings to not
be deselected. At times this caused the HID to remain up and displaying
the event, even though it was removed from the spatial display and the
console.

Fixed bug that caused exceptions to occasionally be thrown when adding
or removing a vertex from a polygon soon after starting Hazard Services.

Change-Id: Ibde2b3c9548c0ec600497db90481c261811b7875