VLab Issue #21676 Fix exception related to end time/duration record
Fixed exception that would occur on machine #1 if an event was issued by
a H.S. client on machine #2, and then the CAVE clock ticked over to a
new minute on machine #1.

Also brought the SessionEventUtilities methods into SessionEventManager,
as they didn't belong in a separate class. Doing this allowed the bug
mentioned above to be more easily fixed.