VLab Issue #29138 Add Latest Hazard Event Version Set to Registry
Added a new way of storing hazard events to the registry (in addition
to the existing history list): the "latest version" set. This is a
set of persisted hazard events, each of which has a different event
ID from the next (unlike the events persisted for history lists, of
which there may be multiple objects with the same event identifier,
showing the hazard event at different points in its lifecycle). This
new method allows hazard events to be persisted to the registry by
H.S. clients (and thus shared among clients connected to the same
registry) at frequent intervals without clogging up the registry with
many copies of the same event.

The history list is still used to save events whenever they are
issued, reissued, ended, or proposed, or when the user manually saves
events via right-click context menu. The latest version persistence
is used by recommenders when they wish to share their changes with
other H.S. clients.

Also pruned the events that are part of a response to a query so that
only those required by the query are included; this means less
serialization and deserialization of events for no purpose as they
are sent from registry to client. Additionally, queries may now be
finer-grained in their approaches, indicating whether they want only
the latest version of a hazard event (either from the latest-version
set or from both that or the history list, whichever is latest); the
history list for a hazard event; or both the latest and the history
list. Also, a query of the size of events' history lists is possible,
which avoids serializing any events at all for the response.

Added ability for recommenders to request that events they generate
or modify either be saved to history lists, saved to the latest
version set, or both (or neither). There are now two event set
attributes that the recommenders can use to signal what should be
saved and how: "saveToHistory" and "saveToDatabase". Recommenders can
indicate they want everything they generate saved to one or the other
category by simply using True as the value for the appropriate event
set attribute, or they can provide a list of one or more event
identifiers for either attribute indicating they want those saved in
the manner associated with that attribute.

Changed the Convective Recommender to take advantage of the new way
of persisting hazard events so that it persists any newly created
events (which are set to status "issued") to history lists, and
persists any modified automated events to the latest version set, so
as to avoid clogging up the history lists of automated events with
numerous versions with minor differences. The Swath Recommender
always saves to the latest version set, not the history list.

Change-Id: I45f9ce212889e8ef5e29943962f2c074b121509f