VLab Issue #29138 Registry notifications and persistence
Changed to ensure that notifications from the registry about added,
modified, or deleted hazard events are processed on the right thread.
Previously, they were processed on arbitrary threads, which led to race
conditions.

Changed persistence of hazard events to registry to include the
"issueTime" attribute if the persistence is the result of issuance or
reissuance, and to not include said attribute if persistence is the
result of user- or recommender-requested saving.

Ensured that pending events that are added to the session as a result of
them appearing in the registry do not get auto-selected.

Added method to hazard event manager get latest hazard events by event
identifier, and fixed improper method name usage in Python calls to the
manager.

Fixed exception caused by recommenders passing back one or more null
elements in the optionally included "saveToDatabase" or "saveToHistory"
lists.

Fixed bugs in ConvectiveRecommender, as well as adding Tracy Hansen's
suggested fix to the Swath Recommender to get it to stop saving to
database hazard events that are merely having their visual features
regenerated. Also changed both Swath and Convective Recommenders to not
return events that have not been changed at all by the recommenders,
especially elapsed ones; and to not save to database or history list
any events that have not changed either. Also added changes made by
Kevin Manross to Convective Recommender while he was at the HWT for
initial testing. Also changed name of generic attribute "status" to
be "statusForHiddenField".

Added switch to StartUpConfig.py allowing the user to configure the
behavior of the persistence requested by recommenders: "normal" persists
events to database or history as before, "history" treats all
persistence as to history, and "none" doesn't persist either.
Persistence when issuing or saving manually still occurs as before; this
switch only changes behavior of recommender-requested persistence.

Added fix allowing the specification of None in a recommender for an
element in a list passed as the result event set's saveToHistory or
saveToDatabase attribute to mean that all newly created events (those
without event identifiers) should be persisted to either history or
database.

Change-Id: Ia9cfd3c7add21625d0e8699cfb994eccfeb1555e