VLab Issue #29138 Fix for registry query, console tree, recommender
Fixed queries made to fetch hazard events from the registry so that they
never include the status as one of their parameters, instead applying
status as a filter to the hazard events returned by the query. This had
to be done because a query asking for events that are not "ended", for
example, would return versions of any event that was not "ended", even
if the current version of said event was indeed "ended". This fixes
problems such as ending an event, then having that event show up as
issued when starting up Hazard Services again.

Fixed console tree visuals bug that caused the time scale widgets in the
rightmost column of any rows below a row that was removed (because for
example it changed statuses to one that is invisible) to not move up by
one row.

Fixed Swath Recommender to not process "ending" status hazard events.
This seems like a good idea anyway, but is needed because switching a
status to "ending" sometimes caused Hazard Services to get caught in a
loop of continuously running the Swath Recommender for the same event,
causing the latter to switch back and for between "issued" and "ending"
endlessly.

Fixed bug in Convective Recommender that meant that events that were
elapsed by said recommender, and were to be returned and saved to the
database/history list, were not returned. The Swath Recommender's
returned event set was erroneously being used as the set that the
Convective Recommender was returning.

Change-Id: I62f755dcd102777ec28bd91b4f12f5379739b74e