VLab Issue #29138 History list and PHI Grid Recommender
Put in code that, using a hazard attribute, indicates which hazards that
are saved to the database are historical versions (i.e. belong in the
history list) and which are not. This will not be needed once the
capability to save-to-latest is restored, but since that causes registry
problems right now, it will need to wait until a later version of the
AWIPS2 baseline is in use with H.S. Experimental.

This means that the history lists as displayed in the console show only
the items that really should be in the history lists. Also, it should be
possible to revert from last saved again.

Added code to the HazardDataAccess Python module allowing the latest
historical version of an event to be fetched from the database, so that
the PHI Grid Recommender can use that, instead of the current version of
the hazard event (which may be in some intermediate state if a user was
modifying it post-issuance).