VLab Issue #15528 Indicate a modified, issued hazard event in Console
Added the tracking of the "modified" status of hazard events, and the
visual indication of modified hazard events that are issued (or ending)
in the Console via the use of bold text for their corresponding rows.

A hazard event is considered modified when one or  more of its first
class fields experience a change in value (with the exception of hazard
status, and a very few other fields), and/or if one or more of the
hazard's attributes that are wrapped in metadata megawidgets experience
a change in value.

A new metadata megawidget property, "doesNotAffectModifyFlag", may be
specified if desired; if set to True, the attribute with which the
megawidget is associated (as identified by its fieldName property) will
not trigger a change in the "modified" status if its value changes.