VLab Issue #32487 Notifications for hazard event field changes
Added new notification to be dispatched when the site identifier,
workstation, or user name of a hazard event changes. This allows the
Console to respond to such changes by updating the appropriate row.

Removed code that set duration to 60 minutes when regenerating Prob_*
metadata.

In the Swath Recommender, commented out code to reset to default motion
vector when Reset Motion Vector (it will now retain the motion vector).
Also added diagnostic print statements for logging.

Change-Id: I5a8bdfcbfba9c9295ab33ecfbea88bc1f5af23fc