VLab Issue #15528 Synchronize event modified flag between H.S. clients
Added "modified" flag, which was previously part of state of hazard
event only locally, to the hazard events that are persisted to the
registry, so that the hazard event modified flag (and thus the bolding
in the console) is consistent between H.S. clients.

Also added ability for a recommender to indicate, if desired, that it
wants any hazard events it saves to the registry (whether in the history
list or the latest-version set) to be treated as an issuance, that is,
to reset the modified flag. Took advantage of this new capability within
the Convective Recommender, so that when it changes issued events, they
will not show up as modified.

Fixed bug with merging of hazard events (when bringing in new versions
of events from the database) that did not include setting the
workstation and user name to the new version's values.

Fixed exception caused by hiding a pending event by filtering out
"Pending" status, for example. The exception was thrown because no check
was made to ensure that a hazard event being redrawn was not filtered
out.

Fixed exception thrown by Swath Recommender when it ran in response to
the user reverting to the last saved event.

Added fix to Swath Recommender to allow vertex to be added to a Prob_*
event. Also added code to change data layer status attribute when
issuing.