VLab Issue #32732 HWT Year 2 week 2 tweaks
Changed product generation for Prob_* events to use a data layer status
of "" (empty string) when issued, instead of "Synced".

Fixed problem with duration being reverted when a new duration was
chosen for an issued Prob_* event, and then the event's geometry was
nudged.

Added code to allow recommenders to specify that the workstation and
user name should not be set for any events modified by said
recommenders, and changed Swath Recommender to use this new capability
when it is only generating visual features, or when it is only
responding to selection state changes.

Fixed history list fetching code in the registry client to not return
any empty history lists, preventing an exception from being thrown
within the client when changing the filters or anything else that
involves loading all events that should be visible given the current
settings.

Change-Id: Ideefba3536cb3338734cf512ed64e39270b924fa