VLab Issue #32732 Move selected time back on data layer update
Added code to the Swath Recommender to reset the selected time back to
where it was prior to the last data layer update upon such an update.
This will cause the selected time to (due to CAVE) move to the latest
data layer time, then after the Swath Recommender runs, move the
selected time back.