VLab Issue #33142 Reset events problem, undefined hazard types problem
Added capability to delete all snapshots of a particular event (that is,
by event identifier) from the registry and by extension from any H.S.
clients connected to that registry. This is then used for resetting
events so that the registry gets cleared out and the hazard events on
the various clients are also all deleted.

Added tracker of identifiers for those events that have been deleted
since the start of the last execution of a recommender. This way, if the
tracker holds event identifiers that are used by events returned by the
recommender, those events are ignored (since it means the recommender
modified them while they were also being deleted via a reset events,
etc.) instead of merged back into the session manager's list of events
being managed.

Added code to print a warning statement if any hazard types defined
within HazardCategories.py are not defined within HazardTypes.py.