VLab Issue #33376 Misc bug fixes for FL.* and Until Further Notice
Fixed problems introduced into CommmonMetaData and the metadata
configuration files for FL.* events a while back (accidental changing of
variable names so that the scripts did not run).

Fixed a bug introduced in the Console refactor that under certain
conditions caused a sort of the Console's hazard event's to run forever,
locking up CAVE.

Fixed bug introduced when time resolution was added that caused "until
further notice" (UFN) values to be rounded to the nearest minute,
instead of being preserved. Since the UFN value is a magic number,
rounding it caused a number of problems. 

Fixed bug in the HID that caused the UFN checkbox to always be disabled,
regardless of whether it should be or not. Also fixed a similar bug that
caused the HID's Propose button to always be disabled.

Fixed the code for changing hazard types so that if UFN was toggled on
before the change, it is turned off for the new hazard type.

Fixed a bug introduced during the Console refactor that caused the UFN
checked context-sensitive menu item to be ignored when toggled by the
user.