VLab Issue #34684 Recommender triggering via event selection changes
Added capability to trigger recommenders as a result of event(s) being
selected or deselected, on a per-hazard-type basis. Currently only
Prob_Severe and Prob_Tornado events trigger the running of any
recommender in such cases (they run the Swath Recommender), and the
Swath Recommender merely prints a message indicating that it was
executed due to selection/deselection, and does nothing more, when it is
so triggered.

This involved changing the event identifier that is passed (as an
attribute of the event set) to recommenders upon execution in certain
cases to be a set of event identifiers, instead of a single one, and
this meant that all the recommenders that reference said attribute
needed to have their code changed to handle the new set, even if they
only expect one event identifier to reside within the set.

Fixed a couple of bugs in CommonMetaData resulting from merge errors
that occurred as part of the #15561 (merging from H.S. Operational)
changesets of the past few weeks. These problems were making it
impossible to generate metadata for FL.* hazard events.