VLab Issue #23056 HID metadata refreshes sometimes need default values
Added a new metadata megawidget definition property,
"overrideOldValues", that allows a hazard event attribute to be marked
as always needing to take the value specified in the megawidget
definition instead of any old value that it had at the moment of the
metadata refresh.

Changed the "Crest to Use" combo box in the FL.* metadata to use this
flag, so that it always starts off with the value provided in the
metadata definition, instead of any old value it may have had prior to
the refresh.