VLab Issue #36555 Updates to the VAA and other related changes following the Arctic Weather Testbed.
Changes were made primarily based on feedback during the testbed, but also have incorporated some changes
arising out of the VAA Design Review with CP.

Major changes were primarily to the VAA megawidget, as well as providing a database type approach for
tracking volcano metadata, numbering, etc., which results in changes to product generation and formatters.

Changes to metadata files are due to centralizing common methods. This check-in also includes the
functionality to create outlook polygons and formatted text for the Convective SIGMET.

Change-Id: Ia3957917e336214f31ed4e775ecb6457d87eb38d