VLab Issue #38072 Improve performance with batching of notifications
Implemented the elements described in the Redmine issue description.
Notifications sent out by the session manager are now batched so that a
number may be accumulated, and in some cases merged with one another so
that less messages are needed.

Recommender requests are also batched, so that multiple requests for a
recommender execution that may be combined are merged together into a
single request, meaning that the recommenders run far less often in the
case of hazard event modification triggering of recommenders.

Fixed bug related to new implementation that became evident when the
user invoked a modifyRecommender Button megawidget, as well as a bug
caused by metadata being fetched before the time range was set upon a
hazard event type change.

Change-Id: Iedeffc8985063e80c150403a7cb26dca5013e0a7