VLab Issue #38506 Python scripts store/retrieve using registry objects
Added the ability for Python scripts to, via the
GenericRegistryObjectDataAccess.py module, store, update, delete, and
retrieve arbitrary registry objects. Such objects are created as
dictionaries in the Python scripts, and converted to Java objects
(complete with searchable elements) before storage in the registry.
Conversely, such objects are converted back to dictionaries for Python
scripts when the latter retrieve them.