VLab Issue #21730 Default type for hazard events
Added "defaultType" configuration option to both StartUpConfig.py and to
settings, allowing the H.S. configuration to be set up so that when a
hazard event is created by the user via drawing on the spatial display,
the event is automatically given a hazard type.

Fixed bugs in recommender execution batching that had been introduced in
the last couple of code reviews, as well as a couple of bugs in the
original code review. These bugs were causing recommenders to be run
more than they needed to be, and in one case, a null pointer exception.
Also fixed a bug causing a lock-up if metadata reloading was triggered
by a recommender's results.

Change-Id: I642e2b564a3718f8403728551aaca083f16666fa