VLab Issue #29138 Save-to-latest uses replace in registry by default
Changed the start-up configuration options to default to using "replace"
within the registry, instead of saving each latest version of a hazard
event to be a separate registry object. This will be tested for a while
to ensure that the old problems with replacement of registry objects is
now gone as as result of moving to 17.1.1. Assuming it is found to not
cause problems anymore, the option will be removed from start-up
configurability, since it was only placed therein as a temporary fix to
allow H.S. to work without performing registry object replacement.