VLab Issue #40923 Changing hazard events from metadata generators
Added code to allow metadata generators to safely change passed-in
hazard events if they so choose. This was done because although hazard
events' generic attributes may be changed implicitly via the
specification of megawidgets in the metadata, first-class fields such as
start and end time, etc. cannot be changed in this manner. Since the
HazardEvent Python class already allowed events to be changed, script
writers were using this capability, unaware that the original design for
metadata generation did not allow for this to be done safely.