VLab Issue #20739 Locking of hazard events
Implemented hazard event locking as per requirements laid out in the
design. This included merging portions of the following changesets from
H.S. IOC:

 Issue  #7623: Implemented Locking of Hazards. (Implemented differently
     from original changeset.)
 Issue #21504: Hazard Locking. (Implemented differently from original
     changeset.)
 Issue #21504 Fixed elapsed error message. Also a few performance fixes.
     and issue with lock timeouts. (Implemented differently from
     original changeset.)
 Issue #21504 Corrections now lock the hazards. (Implemented
     differently from original changeset.)
 Issue #21504 Fixed locking issue when working with more than one
     thread.
 Issue #32733 HazardEventLockHandler no longer ensures events are in the
     registry. Cave is now responsible for checking that the event has
     been persisted before attempting to lock the event. Note this was
     already done in most cases. (Implemented differently from original
     changeset.)
 Issue #32734 Improve performance of hazard locking by moving the orphan
     check to startup of lock manager.
 Issue #33528 Changed lock status indicator on console to E if locked
     for editing.
 Issue #35726 Shutdown lock manager correctly. (Implemented differently
     from original changeset.)
 Issue #35979 HSC's Lock Status now displays Edit instead of E.
     (Implemented differently from original changeset.)
 Issue #29814 Fix HID button state when breaking locks. (Implemented
     differently from original changeset.)

Extensively refactored code base to hide ObservedHazardEvent from
everything besides the session event manager and related classes, as
session events cannot be directly mutable for anything outside of those
classes. This is because the event manager can be trusted to ensure
event locks are acquired, etc. before modifying a session event, but
other classes cannot. The rest of the system now has read-only direct
access to session events, being provided with views that allow the
holder to see underlying changes to said events, but all modifications
must go through the session event manager.

Added attempts to lock whenever a modification of a session event is
proposed.

Console now shows the lock-holder's workstation and user name in its
events table for any event that is locked.

Added new flag to HazardTypes.py that allows a hazard event's metadata
megawidgets to be regenerated anew each time the event is reselected on
a per-hazard-type basis. This is needed to avoid problems with hazard
event types for which selection state determines the metadata megawidget
states, such as Prob_* events, since otherwise when the HID is docked,
the megawidgets are not regenerated when the event is reselected,
whereas when undocked, they are regenerated.

Fixed bug in recommender manager concerning the handling of
recommender-created events. Also fixed bug with regeneration of metadata
upon selection that manifested itself when deselecting an event with no
hazard type.

Fixed bug with setting of editability of all HID metadata megawidgets in
response to a lock or unlock that caused the wrong previous states to
be recorded upon a lock and then to be restored upon an unlock.

Added code to ensure that when a megawidget mutable property change is a
state change, but the change does not actually cause the state of the
megawidget to change, it does not cause any notifications of a change to
go out. This is an efficiency issue generally speaking, but is important
in the context of locking specifically because the notification of
changes that hadn't reallly occurred was causing the session manager to
unnecessarily lock the associated hazard event to attempt to change its
attribute to something the attribute already had as a value. This caused
events to become locked for no good reason.

Fixed bug with setting of hazard event status within session event
manager using the wrong originator, which led to the Swath Recommender
thinking it needed to respond to the change when it did not.

Modified the Swath Recommender to ask H.S. to keep the event it is
working with locked after saving it to the database when appropriate.
Also made it save to history list instead of merely saving the latest
copy when it is changing the hazard event's status.

Added fix to avoid attempting to save a recommender-created event to the
database or history list if the attempt to merge the event with the
session copy fails due to a locking problem, etc.

Fixed error with event identifier display string generation that caused
the truncating of the string to fail when the site identifier was other
than three characters in length (e.g. "National"). Also made the
Prob_WFO settings use the serial number portion of event identifiers
only when displaying events.

Change-Id: I8fbf5cc8f04103dadba5d3c56395afdf8aaf7781