VLab Issue #33428 Toolbar flexibility and enhanced geometry editing
Changed the Console menubar and toolbar so that the Console view itself
decides what ordering to use for the various buttons, etc. that are
contributed by other views. Previously, the other views decided what
ordering to use for their contributions, which was too inflexible for
what users wanted.

Added a new toolbar button, a drop-down that (a) holds images as well as
text in its drop-down menu, and (b) shows the last-chosen item from said
menu as its menu button image, and makes pressing that button invoke
that item from its menu.

Reorganized the existing draw-shape options (draw polygon, draw ellipse,
draw line, etc.) into a single drop-down of the type described above, so
that the options take up less space on the toolbar.

Added a new drop-down to the toolbar that holds the five "enhanced
geometry editing" modes now used. These indicate how a newly drawn
geometry is to be treated with respect to the currently selected hazard
event (if a single event is selected), the new choices being: create a
new event (previously available); replace the event's existing geometry;
union the new geometry with the event's existing geometry (previously
available as a separate toggle); intersect the new geometry with the
event's existing geometry; and subtract the new geometry from the
event's existing geometry.

Removed the old "add new geometry to selected event" button, as it is no
longer needed with the enhanced editing tools. Also removed the two
semi-buggy edit-polygon buttons, as they were clunky to use and like the
other removed button are no longer needed.

Added a new boolean property, "editableUsingGeometryOps", to visual
features, to allow the creator to indicate whether they should be
editable with the new enhanced geometry operation editing modes.

Implemented the new enhanced geometry editing modes themselves. They
operate upon the single selected hazard event, if there is such an
event, and the event's geometry is editable (if a base geometry) or if
it has the abovementioned "editableUsingGeometryOps" flag set to True
(if a visual feature).

Added "issuance count" to hazard events, so that when they are reissued,
a count is kept of how many times they are issued. This may be helpful
in various ways in the future, but the immediate motivation was to allow
recommenders to be triggered by "status" changes event when the status
has not actually changed, but instead the event has been reissued. It
is modifiable from within Python scripts as well, so that for example
the Convective Recommender can set the issuance count higher when it
pseudo-reissues events. (Note that issuance count is incremented each
time an event is issued (including the first time), not merely when it
is reissued.)

Fixed a bug in the Swath Recommender that determined whether visual
features were to be editable or not based upon a variable that itself
was set prior to determining whether the "activate" attribute should be
True or False, which caused editable visual features when they should
have been uneditable.

Fixed a bug in the metadata interdependency script that caused changes
to the "activate" attribute value to not result in the corresponding
changing of editability of appropriate megawidgets.

Change-Id: I5728a488e420c1cf4a85d479b3c9142aeda31ca8